[refactor] rename modal file name to 'models'
